added a results state ,created a new component Results and Meaning
to keep track of the change of results once a key word is searched. Added a component called Results where the result of the search is sent to.
Results -added condiiton if there is no results it returns null, if there is result it returns div with the word searched + loop through the meanings  of each word and it sends resuls to a new component called Meanings.
Meanings returns the part of speech- verb, noun and loop through multiple definitions each word has and it will display all of them.